service_notice/etl/requirements.txt
pydantic[dotenv]==1.10.5
orjson==3.8.7
pika==1.3.1
Jinja2==3.1.2
psycopg2-binary==2.9.5
redis==4.5.1
opentelemetry-api==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-exporter-jaeger==1.16.0
requests==2.28.2
backoff==2.2.1
